Rename for new asm and enhancements
