+++ Code-Samples/Laning-and-Zierler-Interpreter/TestRefs/fl-wwasm.log
@@ -1,0 +1,2 @@
+Corefile output to file las-float-lib.acore
+Listing output to file las-float-lib.lst
