+++ Code-Samples/Laning-and-Zierler-Interpreter/TestRefs/fl-wwsim.log
@@ -1,0 +1,24 @@
+xxxxxxxxxxxx 24 233.081894
+xxxxxxxxxxxx 23 246.941681
+xxxxxxxxxxxx 22 261.625610
+xxxxxxxxxxxx 21 277.182678
+xxxxxxxxxxxx 20 293.664825
+xxxxxxxxxxxx 19 311.127045
+xxxxxxxxxxxx 18 329.627625
+xxxxxxxxxxxx 17 349.228333
+xxxxxxxxxxxx 16 369.994537
+xxxxxxxxxxxx 15 391.995575
+xxxxxxxxxxxx 14 415.304871
+xxxxxxxxxxxx 13 440.000214
+xxxxxxxxxxxx 12 466.164001
+xxxxxxxxxxxx 11 493.883575
+xxxxxxxxxxxx 10 523.251465
+xxxxxxxxxxxx 9 554.365662
+xxxxxxxxxxxx 8 587.329956
+xxxxxxxxxxxx 7 622.254456
+xxxxxxxxxxxx 6 659.255676
+xxxxxxxxxxxx 5 698.457092
+xxxxxxxxxxxx 4 739.989563
+xxxxxxxxxxxx 3 783.991638
+xxxxxxxxxxxx 2 830.610229
+xxxxxxxxxxxx 1 880.000916
