+++ Code-Samples/Nim/nim-fb.ww
@@ -2045,5 +2045,5 @@
 @3577:077565            sp   i3565     ; sub-program 
 
                         .jumpto 0o705     ;  
-                        .ww_file 102766757_fb131-247-12_12-4.7ch/102766757_fb131-247-12_12-4_gs001_fb131-247-12.tcore  ;  
-                        .ww_tapeid _fb131-247-12+_fb  ;  
+                        .ww_file "102766757_fb131-247-12_12-4.7ch/102766757_fb131-247-12_12-4_gs001_fb131-247-12.tcore"  ;  
+                        .ww_tapeid "_fb131-247-12+_fb"  ;
