+++ Code-Samples/Selection-Sort/selection-sort.acore
@@ -1,7 +1,7 @@
 
 ; *** Core Image ***
-%File: ../Code-Samples/Selection-Sort/selection-sort.ww
-%TapeID: None
+%File: selection-sort.ww
+%TapeID: 
 %JumpTo 0o40
 @C0040: 0104200 0040201 0100201 0040202 0100206 0040203 0100204 0044073 
 @C0050: 0040205 0130205 0044054 0044057 0100000 0114203 0070063 0100000 
@@ -23,48 +23,48 @@
 @S0201: count1
 @S0202: count2
 @S0203: max
-@S0204: x1
-@S0205: x2
+@S0204: xi
+@S0205: xj
 @S0206: data
-@E0070: print(f"### max: {py_int(cm.rd(rl('max')))} ###")
-@E0074: for i in range(0, 10): print(f"### data[{i}] = {py_int(cm.rd(rl('data') + i))} ###")
-@N0040: load initial value for outer loop counter
-@N0041: reset outer loop counter
-@N0042: load initial value for inner loop counter
-@N0043: reset inner counter
-@N0044: load x1
-@N0045: x1 is max for now
-@N0046: load address of x1
-@N0047: copy address of x1 into instruction
-@N0050: store address of x1 as address of x2
-@N0051: store address of x1 + 1 as address of x2
-@N0052: copy address of x2 into instruction
-@N0053: copy address of x2 into instruction
-@N0054: load value of x2
-@N0055: compare x2 to max
-@N0056: jump if smaller than max
-@N0057: load value of x2
-@N0060: x2 is the new max
-@N0061: load address of max
-@N0062: copy address of max into instruction
-@N0063: increment x2 address
-@N0064: increment x2 address
-@N0065: increment x2 address
-@N0066: increment inner loop counter
-@N0067: continue inner loop if counter hasn't expired
-@N0070: load value to swap
-@N0071: write value to address of max
-@N0072: load value to swap
-@N0073: write value to address of x1
-@N0074: increment x1 address
-@N0075: increment x1 address
-@N0076: increment x1 address
-@N0077: increment outer loop counter
-@N0100: continue outer loop if counter hasn't expired
-@N0101: halt
-@N0200: (n - 2) * 2^-15 to perform n - 1 comparisons
-@N0201: outer loop counter
-@N0202: inner loop counter
-@N0203: maximum value seen in this inner loop
-@N0204: address of x1
-@N0205: address of x2
+@E0070: exec: wwlog(f"### max: {py_int(cm.rd(rl('max')))} ###")
+@E0074: exec: for i in range(0, 10): print(f"### data[{i}] = {py_int(cm.rd(rl('data') + i))} ###")
+@N0040:  load initial value for outer loop counter
+@N0041:  reset outer loop counter
+@N0042:  load initial value for inner loop counter
+@N0043:  reset inner counter
+@N0044:  load xi
+@N0045:  xi is max for now
+@N0046:  load address of xi
+@N0047:  copy address of xi into instruction
+@N0050:  store address of xi as address of xj
+@N0051:  store address of xi + 1 as address of xj
+@N0052:  copy address of xj into instruction
+@N0053:  copy address of xj into instruction
+@N0054:  load value of xj
+@N0055:  compare xj to max
+@N0056:  jump if smaller than max
+@N0057:  load value of xj
+@N0060:  xj is the new max
+@N0061:  load address of max
+@N0062:  copy address of max into instruction
+@N0063:  increment xj address
+@N0064:  increment xj address
+@N0065:  increment xj address
+@N0066:  increment inner loop counter
+@N0067:  continue inner loop if counter hasn't expired
+@N0070:  load value to swap
+@N0071:  write value to address of max
+@N0072:  load value to swap
+@N0073:  write value to address of xi
+@N0074:  increment xi address
+@N0075:  increment xi address
+@N0076:  increment xi address
+@N0077:  increment outer loop counter
+@N0100:  continue outer loop if counter hasn't expired
+@N0101:  halt
+@N0200:  (n - 2) * 2^-15 to perform n - 1 comparisons
+@N0201:  outer loop counter
+@N0202:  inner loop counter
+@N0203:  maximum value seen in this inner loop
+@N0204:  address of xi
+@N0205:  address of xj
