+++ Code-Samples/Vibrating-String/fb131-97-setup-and-run-Annotated.ww
@@ -38,19 +38,19 @@
 ; Where's this come from? "Warning: Reading Uninitialized Memory at location 0o720, bank 0"
 ; Profile Python -- it's acting like there's a memory leak; the footprint just keeps getting bigger as it runs
 
-                        .ww_file 102663451_fc131-97-56-t2.7ch  ;  
+                        .ww_file "102663451_fc131-97-56-t2.7ch"  ;  
                         ; initial conditions from DCL-090
-                        .switch LeftInterventionReg 0.00100  ; number of points; i/o address 0o336, 0d222 ; was 0o0400
-                        .switch RightInterventionReg 0.40000  ; i/o address 0o337 
-                        .switch FlipFlopPreset02 0.00010  ;  FF#2 Scope Amplitude Adjustment
-                        .switch FlipFlopPreset03 0.70000  ;  FF#3 Elasticity Factor
+                        .switch LeftInterventionReg, 0.00100  ; number of points; i/o address 0o336, 0d222 ; was 0o0400
+                        .switch RightInterventionReg, 0.40000  ; i/o address 0o337 
+                        .switch FlipFlopPreset02, 0.00010  ;  FF#2 Scope Amplitude Adjustment
+                        .switch FlipFlopPreset03, 0.70000  ;  FF#3 Elasticity Factor
                         .org 0         ;  
 @0000:000000            .word 0         ;  @@WrittenBy a0051 a10 a11 ReadBy a2 a3 a4 a5 a6 a7 a8 a9 
 @0001:000001            .word 1         ;  @@ReadBy a0045 a0046 a1 a0077 a0106 a0135 a0177 a0200 
 @0002:000010 scope_amp: .word 0.00010   ; Flip Flop 2 @@ReadBy a0125 
 @0003:070000  elastic:  .word 0.70000   ; Flip Flop 3 @@ReadBy a0120 
 
-                        .ww_file 102766755_fb131-97-6_patterson-t4.7ch  ;  
+                        .ww_file "102766755_fb131-97-6_patterson-t4.7ch"  ;  
                         .org 32        ;  
 @0040:000336        c1: si   222       ; Red ; was FOUR; si 4 clears indicator-light-panel four @@JumpedToBy a0236 
 @0041:000336            si   222       ;  @@JumpedToBy a0237 
@@ -67,8 +67,10 @@
 @0054:130060            ao   d1        ; increment the shift count *instruction* of n_pts by two 
 @0055:130060            ao   d1        ; with the effect of dividing the number by four 
 @0056:100147            ca   s1        ; fetch constant S2 = 0.37760 
-@0057:150150            dv   s2        ; divide contents of AC by npts_norm, result in BR; 0.37760 / 0.4 = almost 1.0 
-@0060:154000        d1: slr  0         ; Red ;Shift by eight bits *and round off* --> 0o400 @@WrittenBy a0052 a0054 a0055 ReadBy a0054 a0055 
+@0057:150150            dv   s2        ; divide contents of AC by npts_norm, result in BR; 0.37760 / 0.4 = almost 1.0
+						.print "xxx8 %ao %bo"
+@0060:154000        d1: slr  0         ; Red ;Shift by eight bits *and round off* --> 0o400 @@WrittenBy a0052 a0054 a0055 ReadBy a0054 a0055
+						.print "xxx9 %ao %bo"
 @0061:040151            ts   s3        ;  
 @0062:100001     oloop: ca   1         ; Red; Outer display loop;  Fetch a One  @@JumpedToBy a0141 
 @0063:040144            ts loop_cnt      ;  T2+1 must contain a loop counter; initialize it to One
@@ -107,9 +109,12 @@
 @0121:114000        a6: su   0         ; subtract the current word in the second data array (x, old) @@WrittenBy a0104 
 @0122:110000        a7: ad   0         ; add current  @@WrittenBy a0074 
 @0123:110000        a8: ad   0         ; add current  @@WrittenBy a0075 
-@0124:040146            ts   t6        ; save the result in temp t6 
+@0124:040146            ts   t6        ; save the result in temp t6
+						.print "xxx12 %ao %bo %d %o", scope_amp, scope_amp
 @0125:144002            mh   scope_amp ;  Scope Amplitude, default 0o10
-@0126:154014            slr  12        ; 
+						.print "xxx10 %ao %bo"
+@0126:154014            slr  12        ;
+						.print "xxx11 %ao %bo"
 @0127:000600            si   384       ; Select Display Points ; the vertical axis setting is in AC[0:9]
 @0130:100000        a9: ca   0         ; Fetch current point in first data array @@WrittenBy a0071 
 @0131:040000       a10: ts   0         ; Save it on the current point of the second array @@WrittenBy a0105 
@@ -134,7 +139,7 @@
                         .org 400       ; start of the data segment
 @0621:000000        r1: .word +0        ; so is this supposed to be 400 or 401??  @@WrittenBy lpstor1 data_strt 
                         ;                         .jumpto c1        ;
-                        .ww_tapeid fc TAPE 131-97-6 PATTERSON  ;  
+                        .ww_tapeid "fc TAPE 131-97-6 PATTERSON"  ;  
 
 
 
@@ -165,14 +170,21 @@
                         ; number of points.  But we seem to have lost the number of points scale factor...
 @0210:150241            dv   npts_norm  ; divide contents of AC by npts_norm, result in BR; 
                         ; i.e., 0.2 / 0.4  -> 0.4 octal, in the B Reg
-@0211:155001            slh  1         ; - puts 0o01 in AC 
-@0212:155006            slh  0         ; ?? what's with this??  (WAS ZERO) 
+						.print "xxx1 %ao %bo"
+@0211:155001            slh  1         ; - puts 0o01 in AC
+						.print "xxx2 %ao %bo"
+@0212:155006            slh  0         ; ?? what's with this??  (WAS ZERO)
+						.print "xxx3 %ao %bo"
 @0213:064240            ex   incr      ; this puts 1 in the incr location, and 400 in the AC 
-@0214:144240            mh   incr      ; 20r ; multiply 0.00001 times 0.00400 = 0.00000 0.01000 
-@0215:155017            slh  15        ; AC=0o400 
+@0214:144240            mh   incr      ; 20r ; multiply 0.00001 times 0.00400 = 0.00000 0.01000
+						.print "xxx4 %ao %bo"
+@0215:155017            slh  15        ; AC=0o400
+						.print "xxx5 %ao %bo"
 @0216:040241            ts   npts_norm  ;  
-@0217:100240            ca   incr      ;  
-@0220:155001            slh  1         ; double the increment value; now it's TWO 
+@0217:100240            ca   incr      ;
+						.print "xxx6 %ao %bo"
+@0220:155001            slh  1         ; double the increment value; now it's TWO
+						.print "xxx7 %ao %bo"
 @0221:040240            ts   incr      ; @25r ; save the increment step size 
 @0222:104241            cs   npts_norm  ;  
 @0223:040242            ts   y_offset  ; initialize the y-offset to the negative of the number of points = -0o400 
@@ -196,4 +208,4 @@
 @0243:040621  data_strt: ts   401       ;  @@ReadBy a0173 
 @0244:040707  data_end: ts   455       ;  @@WrittenBy a0201 ReadBy a0234 
                         .jumpto 120       ;  
-                        .ww_tapeid fc TAPE 131-97-56 PATTERSON  ;  
+                        .ww_tapeid "fc TAPE 131-97-56 PATTERSON"  ;
