--- Tests/asm/TestRefs/inc1.nlst
@@ -1,27 +1,0 @@
-                    .org 0o40
-@0040:100062        ca x
-@0041:110063        ad y
-@0042:040064        ts z
-@0043:000000        si 0
-                    ; .include "inc2.ww"
-@0044:100050        ca a
-@0045:110051        ad b
-@0046:040052        ts c
-@0047:000000        si 0
-@0050:000071     a: .word 57
-@0051:021313     b: .word +0.271828
-@0052:017777     c: .word 0o17777
-                    ; .include "inc3.ww"
-@0053:100057        ca alpha
-@0054:110060        ad beta
-@0055:040061        ts gamma
-@0056:000000        si 0
-@0057:000071 alpha: .word 57
-@0060:010107  beta: .word +0.12718
-@0061:017777 gamma: .word 0o17777
-                    ; end .include "inc3.ww"
-                    ; end .include "inc2.ww"
-@0062:000052     x: .word 42
-@0063:024066     y: .word +0.314159
-@0064:007777     z: .word 0o7777
-
