--- Tests/bounce/TestInfo.yaml
@@ -1,13 +1,0 @@
-
-testType: AsmSim
-baseName: bounce
-asmOptions:
-simOptions: --CycleLimit 7700
-coreOptions:
-verify:
-  - log:
-      name: wwasm
-      filter: Warning|Error
-  - log:
-      name: wwsim
-      filter: ww_draw_point
