+++ Tests/bounce/runtest.sh
@@ -12,7 +12,7 @@ then
 else
 	asm="$PYTHONPATH/../../Py/Assembler/wwasm.py"		# Use quotes since can't resolve backslash yet -- it's needed for file name translation
 	sim="$PYTHONPATH/../../Py/Sim/wwsim.py"
-	rm -f bounce.acore wwsim.log wwasm.log tmp-wwasm.log tmp-ref-wwasm.log tmp-wwsim.log tmp-ref-wwsim.log
+	rm -f bounce.acore bounce.lst wwsim.log wwasm.log tmp-wwasm.log tmp-ref-wwasm.log tmp-wwsim.log tmp-ref-wwsim.log
 	python $asm bounce.ww -o bounce >&wwasm.log
 	python $sim --CycleLimit 7700 bounce.acore >&wwsim.log
 	egrep "Warning|Error" wwasm.log >&tmp-wwasm.log
